Refactor the projects
