+++ Backend/ETLWebApp/Startup.cs
@@ -8,6 +8,7 @@ using ETLLibrary.Database;
 using ETLLibrary.Database.Managers;
 using ETLLibrary.Database.Utils;
 using ETLLibrary.Interfaces;
+using ETLLibrary.Serializers;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.HttpsPolicy;
